requirements.txt
numpy~=2.2.5
matplotlib~=3.10.1
Pillow
